Correct linebreaks for lines ending in dots and length >74
Currently code in quotedPrintableEncode() has code to ensure that a dot
that would end up at the beginning of a line due to soft line breaks,
is properly encoded. This _should_ not be necessary as RFC2821 (4.5.2)
expect the SMTP client to correctly double escape a single dot to two
dots, but Bug #11731 suggests, this might not be the case with all
setups.

Unfortunatly the current code suffers from a miscalculation, where the
line length is calculated based on an already encoded dot at the end of
the line which effectively leads to an off-by-one and an unnecesary
line-break if the line is ending in a dot and is 75 chars long.

Fix the calculation of maximum line length to only consider the
unescaped dot plus the soft line break (=) character only if there's
actually a linebreak needed.

Test added as well.